Refactor modal logic and split add/delete works modules
Extracted add and delete works logic into separate modules (addWorksForm.js, deleteWorks.js) for better modularity and maintainability. Updated modal.js to use these new modules, simplifying the code and improving separation of concerns. No functional changes to the UI, but code is now cleaner and easier to extend.
